refactor(nms): Move fbc_js_core middleware into server
