+++ nms/server/auth/strategies/OrganizationLocalStrategy.js
@@ -14,7 +14,7 @@
  * @format
  */
 
-import type {FBCNMSMiddleWareRequest} from '../../../fbc_js_core/express_middleware';
+import type {FBCNMSMiddleWareRequest} from '../../../server/middleware';
 import type {UserType} from '../../../fbc_js_core/sequelize_models/models/user.js';
 
 import bcrypt from 'bcryptjs';
